refactor(utils): :truck: move rocv_io into the utils package
