+++ rocrate_validator/utils/rocv_io/output/text/layout/report.py
@@ -30,7 +30,7 @@ from rich.text import Text
 import rocrate_validator.log as logging
 from rocrate_validator.colors import get_severity_color
 from rocrate_validator.events import Event, EventType
-from rocrate_validator.rocv_io.output.console import Console
+from rocrate_validator.utils.rocv_io.output.console import Console
 from rocrate_validator.models import (Severity, ValidationContext,
                                       ValidationResult, ValidationSettings,
                                       ValidationStatistics)
